Refactor: 상수 파일 분리 및 추가
