+++ src/constants/racing-rule.js
@@ -1,0 +1,26 @@
+const MAX_RANDOM_NUMBER = 9;
+const MIN_RANDOM_NUMBER = 0;
+
+const MIN_NAME_LENGTH = 1;
+const MAX_NAME_LENGTH = 5;
+
+const MIN_USER = 1;
+const MAX_USER = Number.MAX_SAFE_INTEGER;
+
+const MOVE_STANDARD = 4;
+
+const SKID_MARK = '-';
+
+const DEFAULT_RACING_ROUND = 5;
+
+module.exports = {
+  MAX_RANDOM_NUMBER,
+  MIN_RANDOM_NUMBER,
+  MOVE_STANDARD,
+  MIN_NAME_LENGTH,
+  MAX_NAME_LENGTH,
+  DEFAULT_RACING_ROUND,
+  MIN_USER,
+  MAX_USER,
+  SKID_MARK,
+};
